Get rid of punned casts as well as other compiler warnings. JIRA: CXX-1630
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@45563 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/dbapi/driver/ftds64/freetds/replacements/vasprintf.c
@@ -53,7 +53,7 @@ vasprintf(char **ret, const char *fmt, va_list ap)
 			return -1;
 		}
 		len = vsnprintf(buf, buflen, fmt, ap);
-		if (len >= 0 && len < (buflen - 1)) {
+		if (len >= 0 && len < ((int)buflen - 1)) {
 			break;
 		}
 		free(buf);
@@ -62,7 +62,7 @@ vasprintf(char **ret, const char *fmt, va_list ap)
 		 * len >= 0 are required for vsnprintf implementation that 
 		 * return -1 of buffer insufficient
 		 */
-		if (len >= 0 && len >= buflen) {
+		if (len >= 0 && len >= (int)buflen) {
 			buflen = len + 1;
 		}
 	}
